tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-app-history-back-forward-cross-doc.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  i.contentWindow.appHistory.navigate("/common/blank.html?1");
  await new Promise(resolve => i.onload = resolve);
  assert_equals(i.contentWindow.appHistory.entries().length, 2);

  i.contentWindow.appHistory.oncurrentchange = t.unreached_func("currentchange should not fire for cross-document navigations");
  i.contentWindow.appHistory.back();
  await new Promise(resolve => i.onload = resolve);

  i.contentWindow.appHistory.oncurrentchange = t.unreached_func("currentchange should not fire for cross-document navigations");
  i.contentWindow.appHistory.forward();
  await new Promise(resolve => i.onload = resolve);
}, "AppHistoryCurrentChangeEvent does not fire for cross-document appHistory.back() and appHistory.forward()");
</script>
